Change DamageComponent.DamageMultiplier type from int to float

+++ Monsters.Domain/Monsters/Skills/Components/DamageComponent.cs
@@ -2,6 +2,6 @@
 {
     public class DamageComponent : Component
     {
-        public int DamageMultiplier { get; set; }
+        public float DamageMultiplier { get; set; }
     }
 }
